+++ packages/e2e/ios/Classes/E2EPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface E2EPlugin : NSObject <FlutterPlugin>
+@end