+++ packages/image_picker/example/android/app/src/test/java/io/flutter/plugins/imagepicker/ImagePickerPluginTest.java
@@ -109,6 +109,14 @@ public class ImagePickerPluginTest {
         "No exception thrown when ImagePickerPlugin.registerWith ran with activity = null", true);
   }
 
+  @Test
+  public void onConstructor_WhenContextTypeIsActivity_ShouldNotCrash() {
+    when(mockRegistrar.context()).thenReturn(mockActivity);
+    new ImagePickerPlugin(mockRegistrar, mockImagePickerDelegate);
+    assertTrue(
+        "No exception thrown when ImagePickerPlugin() ran with context instanceof Activity", true);
+  }
+
   private MethodCall buildMethodCall(final int source) {
     final Map<String, Object> arguments = new HashMap<>();
     arguments.put("source", source);