+++ packages/instrumentation_adapter/example/test_driver/widget_test.dart
@@ -1,9 +1,0 @@
-import 'dart:async';
-
-import 'package:flutter_driver/flutter_driver.dart';
-
-Future<void> main() async {
-  final FlutterDriver driver = await FlutterDriver.connect();
-  await driver.requestData(null, timeout: const Duration(minutes: 1));
-  driver.close();
-}